Added new Contact Field (previously messed up the file locations) + tidied up some code from other contributors
