--- EventCard.java
@@ -1,73 +1,0 @@
-package seedu.address.ui;
-
-import javafx.fxml.FXML;
-import javafx.scene.control.Label;
-import javafx.scene.layout.FlowPane;
-import javafx.scene.layout.HBox;
-import javafx.scene.layout.Region;
-import seedu.address.model.event.Event;
-
-/**
- * An UI component that displays information of a {@code Event}.
- */
-public class EventCard extends UiPart<Region> {
-
-    private static final String FXML = "EventListCard.fxml";
-
-    /**
-     * Note: Certain keywords such as "location" and "resources" are reserved keywords in JavaFX.
-     * As a consequence, UI elements' variable names cannot be set to such keywords
-     * or an exception will be thrown by JavaFX during runtime.
-     *
-     * @see <a href="https://github.com/se-edu/addressbook-level4/issues/336">The issue on EventManager level 4</a>
-     */
-
-    public final Event event;
-
-    @FXML
-    private HBox cardPane;
-    @FXML
-    private Label name;
-    @FXML
-    private Label contact;
-    @FXML
-    private Label id;
-    @FXML
-    private Label phone;
-    @FXML
-    private Label address;
-    @FXML
-    private Label email;
-    @FXML
-    private FlowPane tags;
-
-    public EventCard(Event event, int displayedIndex) {
-        super(FXML);
-        this.event = event;
-        id.setText(displayedIndex + ". ");
-        name.setText(event.getName().fullName);
-        contact.setText(event.getContact().fullContactName);
-        phone.setText(event.getPhone().value);
-        address.setText(event.getAddress().value);
-        email.setText(event.getEmail().value);
-        event.getTags().forEach(tag -> tags.getChildren().add(new Label(tag.tagName)));
-    }
-
-    @Override
-    public boolean equals(Object other) {
-        // short circuit if same object
-        if (other == this) {
-            return true;
-        }
-
-        // instanceof handles nulls
-        if (!(other instanceof EventCard)) {
-            return false;
-        }
-
-        // state check
-        EventCard card = (EventCard) other;
-        return id.getText().equals(card.id.getText())
-                && event.equals(card.event);
-    }
-}
