--- PersonCardHandle.java
@@ -1,93 +1,0 @@
-package guitests.guihandles;
-
-import java.util.List;
-import java.util.stream.Collectors;
-
-import com.google.common.collect.ImmutableMultiset;
-
-import javafx.scene.Node;
-import javafx.scene.control.Label;
-import javafx.scene.layout.Region;
-import seedu.address.model.event.Event;
-
-/**
- * Provides a handle to a event card in the event list panel.
- */
-public class PersonCardHandle extends NodeHandle<Node> {
-    private static final String ID_FIELD_ID = "#id";
-    private static final String NAME_FIELD_ID = "#name";
-    private static final String CONTACT_FIELD_ID = "#contact";
-    private static final String ADDRESS_FIELD_ID = "#address";
-    private static final String PHONE_FIELD_ID = "#phone";
-    private static final String EMAIL_FIELD_ID = "#email";
-    private static final String TAGS_FIELD_ID = "#tags";
-
-    private final Label idLabel;
-    private final Label nameLabel;
-    private final Label contactLabel;
-    private final Label addressLabel;
-    private final Label phoneLabel;
-    private final Label emailLabel;
-    private final List<Label> tagLabels;
-
-    public PersonCardHandle(Node cardNode) {
-        super(cardNode);
-
-        idLabel = getChildNode(ID_FIELD_ID);
-        nameLabel = getChildNode(NAME_FIELD_ID);
-        contactLabel = getChildNode(CONTACT_FIELD_ID);
-        addressLabel = getChildNode(ADDRESS_FIELD_ID);
-        phoneLabel = getChildNode(PHONE_FIELD_ID);
-        emailLabel = getChildNode(EMAIL_FIELD_ID);
-
-        Region tagsContainer = getChildNode(TAGS_FIELD_ID);
-        tagLabels = tagsContainer
-                .getChildrenUnmodifiable()
-                .stream()
-                .map(Label.class::cast)
-                .collect(Collectors.toList());
-    }
-
-    public String getId() {
-        return idLabel.getText();
-    }
-
-    public String getName() {
-        return nameLabel.getText();
-    }
-
-    public String getContact() { return contactLabel.getText(); }
-
-    public String getAddress() {
-        return addressLabel.getText();
-    }
-
-    public String getPhone() {
-        return phoneLabel.getText();
-    }
-
-    public String getEmail() {
-        return emailLabel.getText();
-    }
-
-    public List<String> getTags() {
-        return tagLabels
-                .stream()
-                .map(Label::getText)
-                .collect(Collectors.toList());
-    }
-
-    /**
-     * Returns true if this handle contains {@code event}.
-     */
-    public boolean equals(Event event) {
-        return getName().equals(event.getName().fullName)
-                && getContact().equals(event.getContact().fullContactName)
-                && getAddress().equals(event.getAddress().value)
-                && getPhone().equals(event.getPhone().value)
-                && getEmail().equals(event.getEmail().value)
-                && ImmutableMultiset.copyOf(getTags()).equals(ImmutableMultiset.copyOf(event.getTags().stream()
-                        .map(tag -> tag.tagName)
-                        .collect(Collectors.toList())));
-    }
-}
