--- XmlAdaptedEvent.java
@@ -1,143 +1,0 @@
-package seedu.address.storage;
-
-import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Objects;
-import java.util.Set;
-import java.util.stream.Collectors;
-
-import javax.xml.bind.annotation.XmlElement;
-
-import seedu.address.commons.exceptions.IllegalValueException;
-import seedu.address.model.event.*;
-import seedu.address.model.tag.Tag;
-
-/**
- * JAXB-friendly version of the Event.
- */
-public class XmlAdaptedEvent {
-
-    public static final String MISSING_FIELD_MESSAGE_FORMAT = "Event's %s field is missing!";
-
-    @XmlElement(required = true)
-    private String name;
-    @XmlElement(required = true)
-    private String contact;
-    @XmlElement(required = true)
-    private String phone;
-    @XmlElement(required = true)
-    private String email;
-    @XmlElement(required = true)
-    private String address;
-
-    @XmlElement
-    private List<XmlAdaptedTag> tagged = new ArrayList<>();
-
-    /**
-     * Constructs an XmlAdaptedEvent.
-     * This is the no-arg constructor that is required by JAXB.
-     */
-    public XmlAdaptedEvent() {}
-
-    /**
-     * Constructs an {@code XmlAdaptedEvent} with the given event details.
-     */
-    public XmlAdaptedEvent(String name, String contact, String phone, String email, String address, List<XmlAdaptedTag> tagged) {
-        this.name = name;
-        this.contact = contact;
-        this.phone = phone;
-        this.email = email;
-        this.address = address;
-        if (tagged != null) {
-            this.tagged = new ArrayList<>(tagged);
-        }
-    }
-
-    /**
-     * Converts a given Event into this class for JAXB use.
-     *
-     * @param source future changes to this will not affect the created XmlAdaptedEvent
-     */
-    public XmlAdaptedEvent(Event source) {
-        name = source.getName().fullName;
-        contact = source.getContact().fullContactName;
-        phone = source.getPhone().value;
-        email = source.getEmail().value;
-        address = source.getAddress().value;
-        tagged = source.getTags().stream()
-                .map(XmlAdaptedTag::new)
-                .collect(Collectors.toList());
-    }
-
-    /**
-     * Converts this jaxb-friendly adapted event object into the model's Event object.
-     *
-     * @throws IllegalValueException if there were any data constraints violated in the adapted event
-     */
-    public Event toModelType() throws IllegalValueException {
-        final List<Tag> personTags = new ArrayList<>();
-        for (XmlAdaptedTag tag : tagged) {
-            personTags.add(tag.toModelType());
-        }
-
-        if (name == null) {
-            throw new IllegalValueException(String.format(MISSING_FIELD_MESSAGE_FORMAT, Name.class.getSimpleName()));
-        }
-        if (!Name.isValidName(name)) {
-            throw new IllegalValueException(Name.MESSAGE_NAME_CONSTRAINTS);
-        }
-        final Name modelName = new Name(name);
-
-        if (contact == null) {
-            throw new IllegalValueException(String.format(MISSING_FIELD_MESSAGE_FORMAT, Name.class.getSimpleName()));
-        }
-        final Contact modelContact = new Contact(contact);
-
-        if (phone == null) {
-            throw new IllegalValueException(String.format(MISSING_FIELD_MESSAGE_FORMAT, Phone.class.getSimpleName()));
-        }
-        if (!Phone.isValidPhone(phone)) {
-            throw new IllegalValueException(Phone.MESSAGE_PHONE_CONSTRAINTS);
-        }
-        final Phone modelPhone = new Phone(phone);
-
-        if (email == null) {
-            throw new IllegalValueException(String.format(MISSING_FIELD_MESSAGE_FORMAT, Email.class.getSimpleName()));
-        }
-        if (!Email.isValidEmail(email)) {
-            throw new IllegalValueException(Email.MESSAGE_EMAIL_CONSTRAINTS);
-        }
-        final Email modelEmail = new Email(email);
-
-        if (address == null) {
-            throw new IllegalValueException(String.format(MISSING_FIELD_MESSAGE_FORMAT, Address.class.getSimpleName()));
-        }
-        if (!Address.isValidAddress(address)) {
-            throw new IllegalValueException(Address.MESSAGE_ADDRESS_CONSTRAINTS);
-        }
-        final Address modelAddress = new Address(address);
-
-        final Set<Tag> modelTags = new HashSet<>(personTags);
-        return new Event(modelName, modelContact, modelPhone, modelEmail, modelAddress, modelTags);
-    }
-
-    @Override
-    public boolean equals(Object other) {
-        if (other == this) {
-            return true;
-        }
-
-        if (!(other instanceof XmlAdaptedEvent)) {
-            return false;
-        }
-
-        XmlAdaptedEvent otherEvent = (XmlAdaptedEvent) other;
-        return Objects.equals(name, otherEvent.name)
-                && Objects.equals(contact, otherEvent.contact)
-                && Objects.equals(phone, otherEvent.phone)
-                && Objects.equals(email, otherEvent.email)
-                && Objects.equals(address, otherEvent.address)
-                && tagged.equals(otherEvent.tagged);
-    }
-}
