+++ src/main/java/seedu/address/logic/commands/AddCommand.java
@@ -2,6 +2,7 @@ package seedu.address.logic.commands;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_ADDRESS;
+import static seedu.address.logic.parser.CliSyntax.PREFIX_CONTACT;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_EMAIL;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_NAME;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_PHONE;
@@ -11,6 +12,7 @@ import seedu.address.logic.CommandHistory;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.event.Event;
+import seedu.address.storage.Comments;
 
 /**
  * Adds a event to the event manager.
@@ -22,12 +24,14 @@ public class AddCommand extends Command {
     public static final String MESSAGE_USAGE = COMMAND_WORD + ": Adds an event to the event manager. "
             + "Parameters: "
             + PREFIX_NAME + "NAME "
+            + PREFIX_CONTACT + "CONTACT "
             + PREFIX_PHONE + "PHONE "
             + PREFIX_EMAIL + "EMAIL "
             + PREFIX_ADDRESS + "ADDRESS "
             + "[" + PREFIX_TAG + "TAG]...\n"
             + "Example: " + COMMAND_WORD + " "
-            + PREFIX_NAME + "John Doe "
+            + PREFIX_NAME + "New Year's Party "
+            + PREFIX_CONTACT + "Amy Tan "
             + PREFIX_PHONE + "98765432 "
             + PREFIX_EMAIL + "johnd@example.com "
             + PREFIX_ADDRESS + "311, Clementi Ave 2, #02-25 "
@@ -39,6 +43,7 @@ public class AddCommand extends Command {
 
     private final Event toAdd;
 
+    private Comments comments = new Comments();
 
     /**
      * Creates an AddCommand to add the specified {@code Event}
@@ -55,6 +60,7 @@ public class AddCommand extends Command {
         if (model.hasEvent(toAdd)) {
             throw new CommandException(MESSAGE_DUPLICATE_EVENT);
         }
+        comments.createHtml(comments.getFilePath(), comments.getFileName());
         model.addEvent(toAdd);
         model.commitEventManager();
         return new CommandResult(String.format(MESSAGE_SUCCESS, toAdd));
