+++ src/main/java/seedu/address/model/event/Event.java
@@ -10,37 +10,39 @@ import java.util.Set;
 import seedu.address.model.tag.Tag;
 
 /**
- * Represents a Event in the event manager.
+ * Represents an Event in the event manager.
  * Guarantees: details are present and not null, field values are validated, immutable.
  */
 public class Event {
 
     // Identity fields
     private final Name name;
+    private final Contact contact;
     private final Phone phone;
     private final Email email;
 
     // Data fields
     private final Address address;
     private final Set<Tag> tags = new HashSet<>();
-    private final Attendance attendance;
 
     /**
      * Every field must be present and not null.
      */
-    public Event(Name name, Phone phone, Email email, Address address, Attendance attendance, Set<Tag> tags) {
-        requireAllNonNull(name, phone, email, address, tags);
+    public Event(Name name, Contact contact, Phone phone, Email email, Address address, Set<Tag> tags) {
+        requireAllNonNull(name, contact, phone, email, address, tags);
         this.name = name;
+        this.contact = contact;
         this.phone = phone;
         this.email = email;
         this.address = address;
-        this.attendance = attendance;
         this.tags.addAll(tags);
     }
 
     public Name getName() {
         return name;
     }
+
+    public Contact getContact() { return contact; }
 
     public Phone getPhone() {
         return phone;
@@ -52,10 +54,6 @@ public class Event {
 
     public Address getAddress() {
         return address;
-    }
-
-    public Attendance getAttendance() {
-        return attendance;
     }
 
     /**
@@ -77,7 +75,8 @@ public class Event {
 
         return otherEvent != null
                 && otherEvent.getName().equals(getName())
-                && (otherEvent.getPhone().equals(getPhone()) || otherEvent.getEmail().equals(getEmail()));
+                && (otherEvent.getPhone().equals(getPhone()) || otherEvent.getEmail().equals(getEmail())
+                    || otherEvent.getContact().equals(getContact()));
     }
 
     /**
@@ -96,6 +95,7 @@ public class Event {
 
         Event otherEvent = (Event) other;
         return otherEvent.getName().equals(getName())
+                && otherEvent.getContact().equals(getContact())
                 && otherEvent.getPhone().equals(getPhone())
                 && otherEvent.getEmail().equals(getEmail())
                 && otherEvent.getAddress().equals(getAddress())
@@ -105,13 +105,15 @@ public class Event {
     @Override
     public int hashCode() {
         // use this method for custom fields hashing instead of implementing your own
-        return Objects.hash(name, phone, email, address, tags);
+        return Objects.hash(name, contact, phone, email, address, tags);
     }
 
-    @Override
+    @Override //TODO CHECK FOR CONTACT
     public String toString() {
         final StringBuilder builder = new StringBuilder();
         builder.append(getName())
+                .append(" Contact: ")
+                .append(getContact())
                 .append(" Phone: ")
                 .append(getPhone())
                 .append(" Email: ")
