+++ src/main/java/seedu/address/model/util/SampleDataUtil.java
@@ -6,12 +6,8 @@ import java.util.stream.Collectors;
 
 import seedu.address.model.EventManager;
 import seedu.address.model.ReadOnlyEventManager;
+import seedu.address.model.event.*;
 import seedu.address.model.event.Address;
-import seedu.address.model.event.Attendance;
-import seedu.address.model.event.Email;
-import seedu.address.model.event.Event;
-import seedu.address.model.event.Name;
-import seedu.address.model.event.Phone;
 import seedu.address.model.tag.Tag;
 
 /**
@@ -20,23 +16,23 @@ import seedu.address.model.tag.Tag;
 public class SampleDataUtil {
     public static Event[] getSampleEvents() {
         return new Event[] {
-            new Event(new Name("Alex Yeoh"), new Phone("87438807"), new Email("alexyeoh@example.com"),
-                new Address("Blk 30 Geylang Street 29, #06-40"), new Attendance(false),
+            new Event(new Name("Basketball Session"), new Contact("Alex Yeoh"), new Phone("87438807"), new Email("alexyeoh@example.com"),
+                new Address("Blk 30 Geylang Street 29, #06-40"),
                 getTagSet("friends")),
-            new Event(new Name("Bernice Yu"), new Phone("99272758"), new Email("berniceyu@example.com"),
-                new Address("Blk 30 Lorong 3 Serangoon Gardens, #07-18"), new Attendance(false),
+            new Event(new Name("Night Study Session"), new Contact("Bernice Yu"), new Phone("99272758"), new Email("berniceyu@example.com"),
+                new Address("Blk 30 Lorong 3 Serangoon Gardens, #07-18"),
                 getTagSet("colleagues", "friends")),
-            new Event(new Name("Charlotte Oliveiro"), new Phone("93210283"), new Email("charlotte@example.com"),
-                new Address("Blk 11 Ang Mo Kio Street 74, #11-04"), new Attendance(false),
+            new Event(new Name("House Themed Dinner"), new Contact("Charlotte Oliveiro"), new Phone("93210283"), new Email("charlotte@example.com"),
+                new Address("Blk 11 Ang Mo Kio Street 74, #11-04"),
                 getTagSet("neighbours")),
-            new Event(new Name("David Li"), new Phone("91031282"), new Email("lidavid@example.com"),
-                new Address("Blk 436 Serangoon Gardens Street 26, #16-43"), new Attendance(false),
+            new Event(new Name("Cooking Club Tryouts"), new Contact("David Li"), new Phone("91031282"), new Email("lidavid@example.com"),
+                new Address("Blk 436 Serangoon Gardens Street 26, #16-43"),
                 getTagSet("family")),
-            new Event(new Name("Irfan Ibrahim"), new Phone("92492021"), new Email("irfan@example.com"),
-                new Address("Blk 47 Tampines Street 20, #17-35"), new Attendance(false),
+            new Event(new Name("Migrant Workers Rights Sharing Session"), new Contact("Irfan Ibrahim"), new Phone("92492021"), new Email("irfan@example.com"),
+                new Address("Blk 47 Tampines Street 20, #17-35"),
                 getTagSet("classmates")),
-            new Event(new Name("Roy Balakrishnan"), new Phone("92624417"), new Email("royb@example.com"),
-                new Address("Blk 45 Aljunied Street 85, #11-31"), new Attendance(false),
+            new Event(new Name("Ultimate Frisbee Session"), new Contact("Roy Balakrishnan"), new Phone("92624417"), new Email("royb@example.com"),
+                new Address("Blk 45 Aljunied Street 85, #11-31"),
                 getTagSet("colleagues"))
         };
     }
