+++ src/main/java/seedu/address/ui/EventCard.java
@@ -29,6 +29,8 @@ public class EventCard extends UiPart<Region> {
     @FXML
     private Label name;
     @FXML
+    private Label contact;
+    @FXML
     private Label id;
     @FXML
     private Label phone;
@@ -42,8 +44,9 @@ public class EventCard extends UiPart<Region> {
     public EventCard(Event event, int displayedIndex) {
         super(FXML);
         this.event = event;
-        id.setText(displayedIndex + ". ");
+        id.setText(displayedIndex + " ");
         name.setText(event.getName().fullName);
+        contact.setText(event.getContact().fullContactName);
         phone.setText(event.getPhone().value);
         address.setText(event.getAddress().value);
         email.setText(event.getEmail().value);
