+++ src/test/java/seedu/address/testutil/EventBuilder.java
@@ -4,7 +4,7 @@ import java.util.HashSet;
 import java.util.Set;
 
 import seedu.address.model.event.Address;
-import seedu.address.model.event.Attendance;
+import seedu.address.model.event.Contact;
 import seedu.address.model.event.Email;
 import seedu.address.model.event.Event;
 import seedu.address.model.event.Name;
@@ -17,25 +17,26 @@ import seedu.address.model.util.SampleDataUtil;
  */
 public class EventBuilder {
 
-    public static final String DEFAULT_NAME = "Alice Pauline";
+    public static final String DEFAULT_NAME = "Halloween Party";
+    public static final String DEFAULT_CONTACT = "Alice Pauline";
     public static final String DEFAULT_PHONE = "85355255";
     public static final String DEFAULT_EMAIL = "alice@gmail.com";
     public static final String DEFAULT_ADDRESS = "123, Jurong West Ave 6, #08-111";
     public static final Boolean DEFAULT_ATTENDANCE = false;
 
     private Name name;
+    private Contact contact;
     private Phone phone;
     private Email email;
     private Address address;
-    private Attendance attendance;
     private Set<Tag> tags;
 
     public EventBuilder() {
         name = new Name(DEFAULT_NAME);
+        contact = new Contact(DEFAULT_CONTACT);
         phone = new Phone(DEFAULT_PHONE);
         email = new Email(DEFAULT_EMAIL);
         address = new Address(DEFAULT_ADDRESS);
-        attendance = new Attendance(DEFAULT_ATTENDANCE);
         tags = new HashSet<>();
     }
 
@@ -44,10 +45,10 @@ public class EventBuilder {
      */
     public EventBuilder(Event eventToCopy) {
         name = eventToCopy.getName();
+        contact = eventToCopy.getContact();
         phone = eventToCopy.getPhone();
         email = eventToCopy.getEmail();
         address = eventToCopy.getAddress();
-        attendance = eventToCopy.getAttendance();
         tags = new HashSet<>(eventToCopy.getTags());
     }
 
@@ -92,15 +93,15 @@ public class EventBuilder {
     }
 
     /**
-     * Sets the {@code Attendance} of the {@code Event} that we are building.
+     * Sets the {@code Contact} of the {@code Event} that we are building.
      */
-    public EventBuilder withAttendance(String attendance) {
-        this.attendance = new Attendance(attendance);
+    public EventBuilder withContact(String contact) {
+        this.contact = new Contact(contact);
         return this;
     }
 
     public Event build() {
-        return new Event(name, phone, email, address, attendance, tags);
+        return new Event(name, contact, phone, email, address, tags);
     }
 
 }
