+++ src/test/java/seedu/address/testutil/TypicalPersons.java
@@ -23,37 +23,40 @@ import seedu.address.model.event.Event;
  */
 public class TypicalPersons {
 
-    public static final Event ALICE = new EventBuilder().withName("Alice Pauline")
+    public static final Event ALICE = new PersonBuilder().withName("Inter Neighbourhood Cooking Competition")
+            .withContact("Alice Pauline")
             .withAddress("123, Jurong West Ave 6, #08-111").withEmail("alice@example.com")
             .withPhone("94351253")
             .withTags("friends").build();
-    public static final Event BENSON = new EventBuilder().withName("Benson Meier")
+    public static final Event BENSON = new PersonBuilder().withName("Dancing Tryouts")
+            .withContact("Benson Meier")
             .withAddress("311, Clementi Ave 2, #02-25")
             .withEmail("johnd@example.com").withPhone("98765432")
             .withTags("owesMoney", "friends").build();
-    public static final Event CARL = new EventBuilder().withName("Carl Kurz").withPhone("95352563")
-            .withEmail("heinz@example.com").withAddress("wall street").build();
-    public static final Event DANIEL = new EventBuilder().withName("Daniel Meier").withPhone("87652533")
-            .withEmail("cornelia@example.com").withAddress("10th street").withTags("friends").build();
-    public static final Event ELLE = new EventBuilder().withName("Elle Meyer").withPhone("9482224")
-            .withEmail("werner@example.com").withAddress("michegan ave").build();
-    public static final Event FIONA = new EventBuilder().withName("Fiona Kunz").withPhone("9482427")
-            .withEmail("lydia@example.com").withAddress("little tokyo").build();
-    public static final Event GEORGE = new EventBuilder().withName("George Best").withPhone("9482442")
-            .withEmail("anna@example.com").withAddress("4th street").build();
+    public static final Event CARL = new PersonBuilder().withName("Frisbee Competition").withContact("Carl Kurz")
+            .withPhone("95352563").withEmail("heinz@example.com").withAddress("wall street").build();
+    public static final Event DANIEL = new PersonBuilder().withName("Art and Crafts").withContact("Daniel Meier")
+            .withPhone("87652533").withEmail("cornelia@msn.com").withAddress("10th street").withTags("friends").build();
+    public static final Event ELLE = new PersonBuilder().withName("Art Symposium").withContact("Elle Meyer")
+            .withPhone("9482224").withEmail("werner@example.com").withAddress("michegan ave").build();
+    public static final Event FIONA = new PersonBuilder().withName("Dining In The Dark").withContact("Fiona Kunz")
+            .withPhone("9482427").withEmail("lydia@example.com").withAddress("little tokyo").build();
+    public static final Event GEORGE = new PersonBuilder().withName("Inter College Games").withContact("George Best")
+            .withPhone("9482442").withEmail("anna@example.com").withAddress("4th street").build();
 
     // Manually added
-    public static final Event HOON = new EventBuilder().withName("Hoon Meier").withPhone("8482424")
-            .withEmail("stefan@example.com").withAddress("little india").build();
-    public static final Event IDA = new EventBuilder().withName("Ida Mueller").withPhone("8482131")
-            .withEmail("hans@example.com").withAddress("chicago ave").build();
+    public static final Event HOON = new PersonBuilder().withName("Eating Competition").withContact("Hoon Meier")
+            .withPhone("8482424").withEmail("stefan@example.com").withAddress("little india").build();
+    public static final Event IDA = new PersonBuilder().withName("Music Session").withContact("Ida Mueller")
+            .withPhone("8482131").withEmail("hans@example.com").withAddress("chicago ave").build();
 
     // Manually added - Event's details found in {@code CommandTestUtil}
-    public static final Event AMY = new EventBuilder().withName(VALID_NAME_AMY).withPhone(VALID_PHONE_AMY)
-            .withEmail(VALID_EMAIL_AMY).withAddress(VALID_ADDRESS_AMY).withTags(VALID_TAG_FRIEND).build();
-    public static final Event BOB = new EventBuilder().withName(VALID_NAME_BOB).withPhone(VALID_PHONE_BOB)
-            .withEmail(VALID_EMAIL_BOB).withAddress(VALID_ADDRESS_BOB).withTags(VALID_TAG_HUSBAND, VALID_TAG_FRIEND)
-            .build();
+    public static final Event AMY = new PersonBuilder().withName("Night Cycling").withContact(VALID_NAME_AMY)
+            .withPhone(VALID_PHONE_AMY).withEmail(VALID_EMAIL_AMY).withAddress(VALID_ADDRESS_AMY)
+            .withTags(VALID_TAG_FRIEND).build();
+    public static final Event BOB = new PersonBuilder().withName("House Movie Screening").withContact(VALID_NAME_BOB)
+            .withPhone(VALID_PHONE_BOB).withEmail(VALID_EMAIL_BOB).withAddress(VALID_ADDRESS_BOB)
+            .withTags(VALID_TAG_HUSBAND, VALID_TAG_FRIEND).build();
 
     public static final String KEYWORD_MATCHING_MEIER = "Meier"; // A keyword that matches MEIER
